+++ packages/deploy-tooling/src/base/deploy.ts
@@ -171,6 +171,12 @@ async function tryDeploy(archive: Buffer, service: Ui5AbapRepositoryService, con
         }
         logger.error('Deployment has failed.');
         logger.debug(getConfigForLogging(config));
+        if (!config.verbose) {
+            logger.error(
+                'Change logging level to debug your issue\n\t(see examples https://github.com/SAP/open-ux-tools/tree/main/packages/deploy-tooling#configuration-examples)'
+            );
+        }
+
         throw e;
     }
 }